+++ package.json
@@ -16,13 +16,13 @@
     "stacktrace-js": "2.0.2"
   },
   "devDependencies": {
-    "asar": "2.1.0",
     "@testing-library/jest-dom": "5.1.1",
     "@testing-library/react": "9.4.1",
     "@types/jest": "25.1.3",
     "@types/nanoid": "2.1.0",
     "@types/react": "16.9.23",
     "@types/react-dom": "16.9.5",
+    "asar": "2.1.0",
     "eslint": "6.8.0",
     "eslint-config-prettier": "6.10.0",
     "eslint-plugin-prettier": "3.1.2",
@@ -30,6 +30,7 @@
     "jest-canvas-mock": "2.2.0",
     "lint-staged": "10.0.8",
     "node-sass": "4.13.1",
+    "pepjs": "0.5.2",
     "prettier": "1.19.1",
     "rewire": "4.0.1",
     "typescript": "3.8.3"
@@ -92,6 +93,7 @@
     "start": "react-scripts start",
     "test": "npm run test:app",
     "test:app": "react-scripts test --env=jsdom --passWithNoTests",
+    "test:debug": "react-scripts --inspect-brk test --runInBand --no-cache",
     "test:code": "eslint --ignore-path .gitignore --ext .js,.ts,.tsx .",
     "test:other": "npm run prettier -- --list-different"
   },
