+++ src/actions/actionHistory.tsx
@@ -10,33 +10,36 @@ import { KEYS } from "../keys";
 
 const writeData = (
   appState: AppState,
-  data: { elements: ExcalidrawElement[]; appState: AppState } | null,
+  updater: () => { elements: ExcalidrawElement[]; appState: AppState } | null,
 ) => {
-  if (data !== null) {
-    return {
-      elements: data.elements,
-      appState: { ...appState, ...data.appState },
-    };
-  }
-  return {};
-};
-
-const testUndo = (shift: boolean) => (
-  event: KeyboardEvent,
-  appState: AppState,
-) => event[KEYS.META] && /z/i.test(event.key) && event.shiftKey === shift;
-
-export const createUndoAction: (h: SceneHistory) => Action = history => ({
-  name: "undo",
-  perform: (_, appState) =>
+  if (
     [
       appState.multiElement,
       appState.resizingElement,
       appState.editingElement,
       appState.draggingElement,
-    ].every(x => x === null)
-      ? writeData(appState, history.undoOnce())
-      : {},
+    ].some(Boolean)
+  ) {
+    const data = updater();
+
+    return data === null
+      ? {}
+      : {
+          elements: data.elements,
+          appState: { ...appState, ...data.appState },
+        };
+  }
+  return {};
+};
+
+const testUndo = (shift: boolean) => (event: KeyboardEvent) =>
+  event[KEYS.META] && /z/i.test(event.key) && event.shiftKey === shift;
+
+type ActionCreator = (history: SceneHistory) => Action;
+
+export const createUndoAction: ActionCreator = history => ({
+  name: "undo",
+  perform: (_, appState) => writeData(appState, () => history.undoOnce()),
   keyTest: testUndo(false),
   PanelComponent: ({ updateData }) => (
     <ToolButton
@@ -49,17 +52,9 @@ export const createUndoAction: (h: SceneHistory) => Action = history => ({
   commitToHistory: () => false,
 });
 
-export const createRedoAction: (h: SceneHistory) => Action = history => ({
+export const createRedoAction: ActionCreator = history => ({
   name: "redo",
-  perform: (_, appState) =>
-    [
-      appState.multiElement,
-      appState.resizingElement,
-      appState.editingElement,
-      appState.draggingElement,
-    ].every(x => x === null)
-      ? writeData(appState, history.redoOnce())
-      : {},
+  perform: (_, appState) => writeData(appState, () => history.redoOnce()),
   keyTest: testUndo(true),
   PanelComponent: ({ updateData }) => (
     <ToolButton
