+++ src/scene/index.ts
@@ -7,19 +7,9 @@ export {
   getElementsWithinSelection,
   getCommonAttributeOfSelectedElements,
   getSelectedElements,
+  getTargetElement,
 } from "./selection";
-export {
-  exportCanvas,
-  loadFromJSON,
-  loadFromBlob,
-  saveAsJSON,
-  restoreFromLocalStorage,
-  saveToLocalStorage,
-  exportToBackend,
-  importFromBackend,
-  loadScene,
-  calculateScrollCenter,
-} from "./data";
+export { normalizeScroll, calculateScrollCenter } from "./scroll";
 export {
   hasBackground,
   hasStroke,
