+++ src/tests/move.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { render, fireEvent } from "./test-utils";
-import { App } from "../index";
+import { App } from "../components/App";
 import * as Renderer from "../renderer/renderScene";
 
 // Unmount ReactDOM from root
@@ -13,7 +13,9 @@ beforeEach(() => {
   renderScene.mockClear();
 });
 
-describe.skip("move element", () => {
+const { __TEST__: h } = window;
+
+describe("move element", () => {
   it("rectangle", () => {
     const { getByToolName, container } = render(<App />);
     const canvas = container.querySelector("canvas")!;
@@ -27,12 +29,10 @@ describe.skip("move element", () => {
       fireEvent.pointerUp(canvas);
 
       expect(renderScene).toHaveBeenCalledTimes(4);
-      const elements = renderScene.mock.calls[3][0];
-      const selectionElement = renderScene.mock.calls[3][1];
-      expect(selectionElement).toBeNull();
-      expect(elements.length).toEqual(1);
-      expect(elements[0].isSelected).toBeTruthy();
-      expect([elements[0].x, elements[0].y]).toEqual([30, 20]);
+      expect(h.appState.selectionElement).toBeNull();
+      expect(h.elements.length).toEqual(1);
+      expect(h.elements[0].isSelected).toBeTruthy();
+      expect([h.elements[0].x, h.elements[0].y]).toEqual([30, 20]);
 
       renderScene.mockClear();
     }
@@ -42,14 +42,13 @@ describe.skip("move element", () => {
     fireEvent.pointerUp(canvas);
 
     expect(renderScene).toHaveBeenCalledTimes(3);
-    const elements = renderScene.mock.calls[2][0];
-    expect(renderScene.mock.calls[2][1]).toBeNull();
-    expect(elements.length).toEqual(1);
-    expect([elements[0].x, elements[0].y]).toEqual([0, 40]);
+    expect(h.appState.selectionElement).toBeNull();
+    expect(h.elements.length).toEqual(1);
+    expect([h.elements[0].x, h.elements[0].y]).toEqual([0, 40]);
   });
 });
 
-describe.skip("duplicate element on move when ALT is clicked", () => {
+describe("duplicate element on move when ALT is clicked", () => {
   it("rectangle", () => {
     const { getByToolName, container } = render(<App />);
     const canvas = container.querySelector("canvas")!;
@@ -63,12 +62,10 @@ describe.skip("duplicate element on move when ALT is clicked", () => {
       fireEvent.pointerUp(canvas);
 
       expect(renderScene).toHaveBeenCalledTimes(4);
-      const elements = renderScene.mock.calls[3][0];
-      const selectionElement = renderScene.mock.calls[3][1];
-      expect(selectionElement).toBeNull();
-      expect(elements.length).toEqual(1);
-      expect(elements[0].isSelected).toBeTruthy();
-      expect([elements[0].x, elements[0].y]).toEqual([30, 20]);
+      expect(h.appState.selectionElement).toBeNull();
+      expect(h.elements.length).toEqual(1);
+      expect(h.elements[0].isSelected).toBeTruthy();
+      expect([h.elements[0].x, h.elements[0].y]).toEqual([30, 20]);
 
       renderScene.mockClear();
     }
@@ -78,11 +75,10 @@ describe.skip("duplicate element on move when ALT is clicked", () => {
     fireEvent.pointerUp(canvas);
 
     expect(renderScene).toHaveBeenCalledTimes(3);
-    const elements = renderScene.mock.calls[2][0];
-    expect(renderScene.mock.calls[2][1]).toBeNull();
-    expect(elements.length).toEqual(2);
+    expect(h.appState.selectionElement).toBeNull();
+    expect(h.elements.length).toEqual(2);
     // previous element should stay intact
-    expect([elements[0].x, elements[0].y]).toEqual([30, 20]);
-    expect([elements[1].x, elements[1].y]).toEqual([0, 40]);
+    expect([h.elements[0].x, h.elements[0].y]).toEqual([30, 20]);
+    expect([h.elements[1].x, h.elements[1].y]).toEqual([0, 40]);
   });
 });
